Ensure that the SELinux file context of the newly installed fonts is correct
When the fonts are downloaded and unpacked in /tmp, they are
automatically labeled with dpkg_script_tmp_t. If nothing is done, that
label is kept when the files are moved to their final location. By
adding the "-Z" option, mv will query the "file context database" and set
the correct label on the file.

This requires coreutils >= 8.22 which is already in debian
stable/stretch

Closes: #899373